backend/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
pydantic>=2.6
pydantic-settings>=2.2
sqlalchemy>=2.0
psycopg[binary]>=3.1
alembic>=1.13
PyJWT>=2.8
